requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery
django-celery-beat
django-celery-results
pytz